Table_Service/auth_service/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
PyJWT
aiokafka